requirements.txt
uvicorn
fastapi
sqlalchemy
python-dotenv
psycopg2-binary